requirements.txt
Flask==1.0.2
findspark==1.3.0
pyspark==2.4.4
PyYAML==5.1.2
